+++ reviewblitz/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
-from reviewblitz.models import ReviewBlitzScoring, ReviewBlitz, Review, BlitzReview
+from reviewblitz.models import ReviewBlitzScoring, ReviewBlitz, BlitzReview
 
 admin.site.register(ReviewBlitzScoring)
 admin.site.register(ReviewBlitz)
-admin.site.register(Review)
 admin.site.register(BlitzReview)
